pyproject.toml
[tool.poetry]
name = "Della Parser Bot"
version = "0.1.0"
description = ""
authors = ["Georgiy Kozhevnikov <baraltiva@gmail.com>"]
packages = [
    { include = "della_parser_bot"}
]
[tool.poetry.group.dev.dependencies]
types-redis = "^4.5.1.5"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
pytelegrambotapi = "^4.10.0"
loguru = "^0.6.0"
types-requests = "^2.28.11.15"
beautifulsoup4 = "^4.11.2"
peewee = "^3.16.0"
python-dotenv = "^1.0.0"
blue = {extras = ["jupyter"], version = "^0.9.1"}
celery = "^5.2.7"
celery-types = "^0.14.0"
redis = "^4.5.1"

[tool.poetry.dev-dependencies]
blue = "*"
mypy = "*"
pip-audit = "*"
pytest = "*"
pytest-cov = "*"
ruff = "*"

[tool.blue]
line-length = 79

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
strict = true
exclude = ["tests/*", "della_parser_bot/src/messages.py"]

[tool.ruff]
line-length = 79
select = [
    "A",
    "ARG",
    "B",
    "C",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "T10",
    "UP",
    "W",
]
ignore = ["A003"]
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "S", "PLR2004", "E501"]
"della_parser_bot/src/messages.py" = ["E501"]
